news_summarizer/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.6
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.1
streamlit>=1.30.0

